requirements.txt
opengradient
flask
flask-cors
python-dotenv
